requirements.txt
PyYAML==3.11
datadog==0.10.0
pytz==2015.7

